+++ include/trade.h
@@ -7,11 +7,11 @@ struct Trade
 {
     int buy_order_id;
     int sell_order_id;
-    double price;
+    int price;
     int quantity;
     long timestamp;
 
-    Trade(int buyId, int sellId, double price, int quantity, long timestamp) : buy_order_id(buyId), sell_order_id(sellId), price(price), quantity(quantity), timestamp(timestamp)
+    Trade(int buyId, int sellId, int price, int quantity, long timestamp) : buy_order_id(buyId), sell_order_id(sellId), price(price), quantity(quantity), timestamp(timestamp)
     {
         if (quantity <= 0)
             throw std::invalid_argument("Quantity must be positive.");
